WideToUtf8 fixes.

+++ src/common/Text/GTextView4.cpp
@@ -1307,7 +1307,7 @@ public:
 			{
 				Text *t = Txt[i];
 				int TxtOffset = BlkOffset - Pos;
-				if (TxtOffset >= 0 && TxtOffset < t->Length())
+				if (TxtOffset >= 0 && TxtOffset < (int)t->Length())
 				{
 					int MaxChars = t->Length() - TxtOffset;
 					int Remove = min(Chars, MaxChars);
